Move path resolution types to their module
