+++ compiler/noirc_frontend/src/elaborator/mod.rs
@@ -4,9 +4,8 @@ use std::{
 };
 
 use crate::{
-    ast::ItemVisibility, hir::resolution::import::PathResolutionItem,
-    hir_def::traits::ResolvedTraitBound, usage_tracker::UsageTracker, StructField, StructType,
-    TypeBindings,
+    ast::ItemVisibility, hir_def::traits::ResolvedTraitBound, usage_tracker::UsageTracker,
+    StructField, StructType, TypeBindings,
 };
 use crate::{
     ast::{
@@ -24,7 +23,6 @@ use crate::{
         def_map::{DefMaps, ModuleData},
         def_map::{LocalModuleId, ModuleId, MAIN_FUNCTION},
         resolution::errors::ResolverError,
-        resolution::import::PathResolution,
         scope::ScopeForest as GenericScopeForest,
         type_check::{generics::TraitGenerics, TypeCheckError},
         Context,
@@ -59,6 +57,7 @@ mod unquote;
 use fm::FileId;
 use iter_extended::vecmap;
 use noirc_errors::{Location, Span, Spanned};
+use path_resolution::{PathResolution, PathResolutionItem};
 use types::bind_ordered_generics;
 
 use self::traits::check_trait_impl_method_matches_declaration;
