+++ compiler/noirc_frontend/src/elaborator/patterns.rs
@@ -9,7 +9,7 @@ use crate::{
     },
     hir::{
         def_collector::dc_crate::CompilationError,
-        resolution::{errors::ResolverError, import::PathResolutionItem},
+        resolution::errors::ResolverError,
         type_check::{Source, TypeCheckError},
     },
     hir_def::{
@@ -20,7 +20,7 @@ use crate::{
     Kind, Shared, StructType, Type, TypeAlias, TypeBindings,
 };
 
-use super::{Elaborator, ResolverMeta};
+use super::{path_resolution::PathResolutionItem, Elaborator, ResolverMeta};
 
 impl<'context> Elaborator<'context> {
     pub(super) fn elaborate_pattern(
